pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kb-agent"
version = "0.1.0"
description = "Enterprise-grade AI Agent for internal bank document retrieval"
authors = [
    {name = "Jules", email = "jules@example.com"},
]
requires-python = ">=3.10"
dependencies = [
    "textual>=0.40.0",
    "chromadb>=0.4.0",
    "ripgrep-python>=0.1.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "rich>=13.0.0",
    "openai>=1.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "tabulate>=0.9.0",
    "requests>=2.31.0",
    "networkx>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.13.0",
    "langgraph>=1.0.0",
    "langchain-openai>=0.3.0",
    "crawl4ai>=0.6.0",
    "rank-bm25>=0.2.0",
    "pymupdf>=1.27.1",
]

[project.scripts]
kb-agent = "kb_agent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
